# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Film-collection API",
    "contact": {
      "name": "Yuliia Kopysova",
      "email": "julykopysova97@b.com"
    }
  },
  "host": "172.20.0.3:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/signup": {
      "post": {
        "summary": "API Endpoint for create User",
        "description": "Create new User in database",
        "tags": [
                    "Users"
                ],
        "parameters": [
          {
            "name": "user",
            "description": "User params to register new user",
            "in": "body",
            "schema" : {

                }
          }
        ],
        "responses": {
          "200": {
            "description": "New user has Created"
          },
          "405": {
            "description": "This data already exists"
          },
          "401": {
            "description": "Incorrect data"
          }
        }
      }
    },
    "/login": {
        "post": {
          "summary": "API Endpoint for log in  User",
          "description": "For log in user in system",
          "tags": [
                      "Users"
                  ],
          "parameters": [
            {
              "name": "user",
              "description": "User enter username and password",
              "in": "body",
              "schema" : {

                  }
            }
          ],
          "responses": {
            "202": {
              "description": "Log In"
            },
            "401": {
              "description": "Username or Password Error"
            }
          }
        }
    },
     "/directors": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get directors",
                "description": ":return: Response",
                "tags": [
                    "Directors"
                ]
            },
            "post": {
                "summary": "API Endpoint for create Director by admin",
                "description": "Create new Director by admin",
                "tags": [
                            "Directors"
                        ],
                "parameters": [
                  {
                    "name": "director",
                    "description": "Director params for director",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "New director has created"
                  },
                  "401": {
                    "description": "User is not admin"
                  }
                }
            }
        },
    "/logout": {
        "post": {
          "summary": "API Endpoint for log out  User",
          "description": "For log out user in system",
          "tags": [
                      "Users"
                  ],
          "responses": {
            "200": {
              "description": "Logout success"
            },
            "401": {
                        "description": "Not authorized"
            }
          }
        }
    },
    "/directors/{director_id}": {
            "parameters": [
                {
                    "name": "director_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "User is not admin"
                    }
                },
                "summary": "Delete director",
                "description": ":param director_id: id of director\n:return: Response",
                "tags": [
                    "Directors"
                ]
            },
            "patch": {
              "parameters": [
                  {
                    "name": "director",
                    "description": "Director params for director",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "summary": "Updates a director",
                "description": ":param director_id: id of director\n:return: JSON",
                "tags": [
                    "Directors"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get one director",
                "description": ":param director_id: id of director\n:return: JSON",
                "tags": [
                    "Directors"
                ]
            }
        },
...
  }
}